Refactor GiphySharingPreviewFragment. Move some code to the SyncEngine.
